Update B_Uniqueness.java
First pass, go through the list building a counting hash map — keys are the integers, values are the frequency in the array. If no duplicates, we’re done, the answer is 0.

Second pass, start with (L,R) = (1,1); at each step, update the counting hash map to reflect the new set of counts if we remove everything from L to R inclusive. If the resulting hash map reflects duplicate values still present, increment R. If the resulting hash map reflects a clean set with no duplicates, increment L. While you’re doing this, keep track of the smallest difference (R-L) which gives a clean set with no duplicates. Work out how to handle it once R hits the end of the array. To know if the hash map has no duplicates, don’t iterate it each time; instead, keep an auxiliary variable which counts how many values > 1 are in the hash map and keep it updated accurately every time a value crosses the 1<->2 threshold.

That should be O(N), assuming that hash map operations are O(1). The first pass should be obviously O(N), and the second pass can only execute roughly 2N iterations, since L can be incremented N-1 times and R can be incremented the same number of times.